// src/modules/fichas/dto/create-ficha.dto.ts
export class CreateFichaDto {
    nombreArchivo: string;

    codigoArchivo: string;

    id_programa: string;

    error: string;

    referencia: string;

    fechaRealizacion: string;

    fechaEmision: string;

    casaProductora: string;

    formato: string;

    tipoArchivoRecibido: string;

    tipoArchivoGuardado: string;

    soporteFisicoGrabacion: string;

    resolucion: string;

    tasaRecuadro: string;

    codecVideo: string;

    codecAudio: string;

    canalesAudio: string;

    tasaMuestra: string;

    bitsPorMuestra: string;

    sistemaAcceso: string;

    duracion: string;

    sinopsis: string;

    observaciones: string;

    copyright: string;
}